Successfull login and register
users need to be added to the spotify api developer account and then they are able to connect to their spotify accounts. Index.js will then check if the user already has an existing account in the database  and if so will navigate them to the Main screen. If the user does not have an existing account in the database, then it will add them using spotify email and username and then navigate them to the Main page.