fixed login/register issues
reformatted the backend to better handle communicating to the database. A token is now generated for a successful login/register which is then used to reference to the database to keep track of the logged in user. Temp profile screen is also updated to pull the info from the database